chore: move val helper (#1060)
* moved validator_helper to model_utils


* changed doc string


* docs validate


* removed unneeded TG
