--- docs/api_reference/trestle.core.validator_helper.md
@@ -1,2 +1,0 @@
-::: trestle.core.validator_helper
-handler: python
