+++ trestle/core/commands/author/profile.py
@@ -34,7 +34,6 @@ from trestle.core.commands.author.common import AuthorCommonCommand
 from trestle.core.commands.common.return_codes import CmdReturnCodes
 from trestle.core.control_io import ParameterRep
 from trestle.core.profile_resolver import ProfileResolver
-from trestle.core.validator_helper import regenerate_uuids
 
 logger = logging.getLogger(__name__)
 
@@ -362,7 +361,7 @@ class ProfileAssemble(AuthorCommonCommand):
                 return CmdReturnCodes.SUCCESS.value
 
         if regenerate:
-            new_profile, _, _ = regenerate_uuids(new_profile)
+            new_profile, _, _ = ModelUtils.regenerate_uuids(new_profile)
         ModelUtils.update_last_modified(new_profile)
 
         if new_profile_dir.exists():
