+++ trestle/core/commands/replicate.py
@@ -20,7 +20,6 @@ import logging
 from trestle.common import const, file_utils, log
 from trestle.common.err import TrestleError, handle_generic_command_exception
 from trestle.common.model_utils import ModelUtils
-from trestle.core import validator_helper
 from trestle.core.commands.command_docs import CommandPlusDocs
 from trestle.core.commands.common.return_codes import CmdReturnCodes
 from trestle.core.models.actions import CreatePathAction, WriteFileAction
@@ -95,7 +94,7 @@ class ReplicateCmd(CommandPlusDocs):
 
         if args.regenerate:
             logger.debug(f'regenerating uuids for model {input_file}')
-            model_instance, uuid_lut, n_refs_updated = validator_helper.regenerate_uuids(model_instance)
+            model_instance, uuid_lut, n_refs_updated = ModelUtils.regenerate_uuids(model_instance)
             logger.debug(f'{len(uuid_lut)} uuids generated and {n_refs_updated} references updated')
 
         # 4 Prepare actions and plan
